pyproject.toml
[tool.poetry]
name = "fastapi-with-graphql"
version = "0.1.0"
description = ""
authors = ["Mark Stella <mark.p.stella@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["fastapi"], version = "^0.229.1"}
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pymongo = "^4.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"